Remove `MemorySpace`
